update ui widgets
