+++ lib/widgets/styled/input_adder/tag.dart
@@ -1,0 +1,27 @@
+import 'package:Organiser/widgets/shared/container/add_tag.dart';
+import 'package:Organiser/widgets/shared/input/buttons.dart';
+import 'package:flutter/material.dart';
+
+// ignore: must_be_immutable
+class TagsAdder extends StatelessWidget {
+  List<Tag> selectedTags = [];
+
+  TagsAdder({required this.selectedTags});
+
+  @override
+  Widget build(BuildContext context) {
+    return StyledButtons.primaryElevatedButton(
+        onPressed: () async {
+          List<Tag>? result = await AddTagDialog.show(context);
+          setState(() {
+            selectedTags = result!;
+          });
+        },
+        verticalPadding: 0,
+        text: 'Tags',
+        icon: Icons.bookmark_add_outlined,
+        context: context);
+  }
+  
+  void setState(Null Function() param0) {}
+}
