+++ lib/widgets/styled/input_adder/title.dart
@@ -1,0 +1,18 @@
+import 'package:Organiser/widgets/shared/input/textfields.dart';
+import 'package:flutter/material.dart';
+
+class TitleInput extends StatelessWidget {
+  final TextEditingController titleController;
+
+  TitleInput({required this.titleController});
+
+  @override
+  Widget build(BuildContext context) {
+    return StyledTextField(
+      controller: titleController,
+      label: 'Title',
+      textSize: 25,
+      trailingIcon: Icons.sports_kabaddi_rounded,
+    );
+  }
+}
